+++ src/Storage/Entity/Invitation.php
@@ -1,0 +1,81 @@
+<?php
+namespace Bolt\Storage\Entity;
+
+/**
+ * Entity for Invitations.
+ */
+class Invitation extends Entity
+{
+    /** @var int */
+    protected $ownerid;
+    /** @var string */
+    protected $token;
+    /** @var \DateTime */
+    protected $expiration;
+    /** @var array */
+    protected $roles = [];
+
+    /**
+     * @return int
+     */
+    public function getOwnerid()
+    {
+        return $this->ownerid;
+    }
+
+    /**
+     * @param int $ownerid
+     */
+    public function setOwnerid($ownerid)
+    {
+        $this->ownerid = $ownerid;
+    }
+
+    /**
+     * @return string
+     */
+    public function getToken()
+    {
+        return $this->token;
+    }
+
+    /**
+     * @param string $token
+     */
+    public function setToken($token)
+    {
+        $this->token = $token;
+    }
+
+    /**
+     * @return \DateTime
+     */
+    public function getExpiration()
+    {
+        return $this->expiration;
+    }
+
+    /**
+     * @param \DateTime $expiration
+     */
+    public function setExpiration($expiration)
+    {
+        $this->expiration = $expiration;
+    }
+
+    /**
+     * @return array
+     */
+    public function getRoles()
+    {
+        return $this->roles;
+    }
+
+    /**
+     * @param array $roles
+     */
+    public function setRoles(array $roles)
+    {
+        $this->roles = array_values(array_unique($roles));
+    }
+}